requirements.txt
flask==3.0.2
# OpenTelemetry dependencies (commented for future use)
# opentelemetry-api==1.23.0
# opentelemetry-sdk==1.23.0
# opentelemetry-instrumentation-flask==0.44b0
# opentelemetry-exporter-otlp==1.23.0
prometheus-client==0.19.0
python-dotenv==1.0.1
google-cloud-monitoring==2.19.0 